backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
python-whois==0.8.0
dnspython==2.3.0
requests==2.30.0
python-dotenv==1.0.0
pydantic==1.10.7
reportlab==3.6.13
weasyprint==59.0
vt-py==0.17.5
openai==1.13.3
aiofiles==23.1.0
pillow==9.5.0
httpx==0.27.0
